Много изменений

+++ app/src/main/java/ru/melodin/fast/CreateChatActivity.kt
@@ -3,6 +3,7 @@ package ru.melodin.fast
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
+import android.view.MenuItem
 import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
@@ -16,15 +17,14 @@ import ru.melodin.fast.adapter.RecyclerAdapter
 import ru.melodin.fast.api.UserConfig
 import ru.melodin.fast.api.VKApi
 import ru.melodin.fast.api.model.VKUser
+import ru.melodin.fast.common.TaskManager
 import ru.melodin.fast.common.ThemeManager
-import ru.melodin.fast.concurrent.AsyncCallback
-import ru.melodin.fast.concurrent.ThreadExecutor
 import ru.melodin.fast.database.CacheStorage
 import ru.melodin.fast.database.DatabaseHelper
 import ru.melodin.fast.util.ArrayUtil
-import ru.melodin.fast.util.Requests
 import ru.melodin.fast.util.Util
 import ru.melodin.fast.util.ViewUtil
+import ru.melodin.fast.view.FastToolbar
 import java.util.*
 
 class CreateChatActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshListener, RecyclerAdapter.OnItemClickListener {
@@ -37,7 +37,7 @@ class CreateChatActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshList
         loadFriends()
     }
 
-    override fun onItemClick(v: View, position: Int) {
+    override fun onItemClick(position: Int) {
         adapter!!.toggleSelected(position)
         adapter!!.notifyItemChanged(position, -1)
         setTitle()
@@ -46,7 +46,7 @@ class CreateChatActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshList
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        setTheme(ThemeManager.getCurrentTheme())
+        setTheme(ThemeManager.currentTheme)
         ViewUtil.applyWindowStyles(window)
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_create_chat)
@@ -54,16 +54,19 @@ class CreateChatActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshList
         setTitle()
 
         tb.setBackVisible(true)
-        tb.setOnBackClickListener { onBackPressed() }
+        tb.setOnBackClickListener(View.OnClickListener { onBackPressed() })
         tb.inflateMenu(R.menu.activity_create_chat)
-        tb.setOnMenuItemClickListener { item ->
-            if (item.itemId == R.id.create)
-                getUsers()
-        }
+        tb.setOnMenuItemClickListener(object : FastToolbar.OnMenuItemClickListener {
+            override fun onMenuItemClick(item: MenuItem) {
+                if (item.itemId == R.id.create)
+                    getUsers()
+            }
+
+        })
 
         refresh.setOnRefreshListener(this)
-        refresh.setColorSchemeColors(ThemeManager.getAccent())
-        refresh.setProgressBackgroundColorSchemeColor(ThemeManager.getBackground())
+        refresh.setColorSchemeColors(ThemeManager.accent)
+        refresh.setProgressBackgroundColorSchemeColor(ThemeManager.background)
 
         val manager = LinearLayoutManager(this)
         manager.orientation = RecyclerView.VERTICAL
@@ -119,32 +122,33 @@ class CreateChatActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshList
         setTitle()
 
         refresh.isRefreshing = true
-        ThreadExecutor.execute(object : AsyncCallback(this) {
+        TaskManager.execute {
 
             lateinit var friends: ArrayList<VKUser>
 
-            @Throws(Exception::class)
-            override fun ready() {
-                friends = VKApi.friends().get().userId(UserConfig.userId).order("hints").fields(VKUser.FIELDS_DEFAULT).execute(VKUser::class.java)
+            VKApi.friends().get().userId(UserConfig.userId).order("hints").fields(VKUser.FIELDS_DEFAULT).execute(VKUser::class.java, object : VKApi.OnResponseListener {
+                override fun onSuccess(models: ArrayList<*>?) {
+                    if (ArrayUtil.isEmpty(models)) return
+                    models?: return
 
-                CacheStorage.delete(DatabaseHelper.FRIENDS_TABLE)
-                CacheStorage.insert(DatabaseHelper.FRIENDS_TABLE, friends)
-                CacheStorage.insert(DatabaseHelper.USERS_TABLE, friends)
-            }
+                    friends = models as ArrayList<VKUser>
+                    CacheStorage.delete(DatabaseHelper.FRIENDS_TABLE)
+                    CacheStorage.insert(DatabaseHelper.FRIENDS_TABLE, friends)
+                    CacheStorage.insert(DatabaseHelper.USERS_TABLE, friends)
 
-            override fun done() {
-                createAdapter(friends)
-                refresh.isRefreshing = false
+                    createAdapter(friends)
+                    refresh.isRefreshing = false
 
-                setTitle()
-            }
+                    setTitle()
+                }
 
-            override fun error(e: Exception) {
-                setTitle()
-                refresh.isRefreshing = false
-                Toast.makeText(this@CreateChatActivity, R.string.error, Toast.LENGTH_LONG).show()
-            }
-        })
+                override fun onError(e: Exception) {
+                    setTitle()
+                    refresh.isRefreshing = false
+                    Toast.makeText(this@CreateChatActivity, R.string.error, Toast.LENGTH_LONG).show()
+                }
+            })
+        }
     }
 
     private fun setTitle() {
@@ -170,12 +174,12 @@ class CreateChatActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshList
 
     private fun createChat(users: ArrayList<VKUser>) {
         val user = UserConfig.getUser()
-        users.add(0, user)
+        users.add(0, user!!)
 
         val intent = Intent(this, ShowCreateChatActivity::class.java)
         intent.putExtra("users", users)
 
-        startActivityForResult(intent, Requests.CREATE_CHAT)
+        startActivityForResult(intent, REQUEST_CREATE_CHAT)
 
         adapter!!.clearSelect()
         adapter!!.notifyItemRangeChanged(0, adapter!!.itemCount, -1)
@@ -194,9 +198,9 @@ class CreateChatActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshList
     }
 
     public override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (requestCode == Requests.CREATE_CHAT) {
+        if (requestCode == REQUEST_CREATE_CHAT) {
             if (resultCode == Activity.RESULT_OK) {
-                data?: return
+                data ?: return
                 val title = data.getStringExtra("title")
                 val peerId = data.getIntExtra("peer_id", -1)
 
@@ -204,5 +208,11 @@ class CreateChatActivity : AppCompatActivity(), SwipeRefreshLayout.OnRefreshList
             }
         }
         super.onActivityResult(requestCode, resultCode, data)
+    }
+
+    companion object {
+
+        const val REQUEST_CREATE_CHAT = 1
+
     }
 }
